labProgress05
-experimented with school grade level cleaning functions as mentioned
couple commits ago.
-- issue: if we were assigning HAS_SCHOOLLEVEL keys to each object based
on what substrings are found in GRADE_LEVEL key, I wasn't sure why we
were checking data type of GRADE_ORG. If GRADE_ORG is the one that is
supposed to have string values, but by entry error has numbers in some
objects, that shouldn't matter to GRADE_LEVEL field we were using to
clean.
-- result: There were 5 objects found that had number values for
GRADE_ORG. When replaced it with another function that solely uses
substring value of GRADE_LEVEL and I got an error. Turns out, out of the
5 that had entry errors for GRADE_ORG, three of them ALSO had entry
errors for GRADE_LEVEL. Essentially what we are concerned for is these
three, not all five. Kept the data type check for cleaning, but changed
GRADE_ORG to GRADE_LEVEL. Cleaning result remained the same.